+++ src/Shared/SpanHelper.cs
@@ -11,29 +11,8 @@ namespace OpenTelemetry.Trace;
 internal static class SpanHelper
 {
     /// <summary>
-    /// Helper method that populates span properties from http status code according
-    /// to https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/http.md#status.
-    /// </summary>
-    /// <param name="httpStatusCode">Http status code.</param>
-    /// <returns>Resolved span <see cref="Status"/> for the Http status code.</returns>
-    public static ActivityStatusCode ResolveSpanStatusForHttpStatusCode(int httpStatusCode)
-    {
-        if (httpStatusCode >= 100 && httpStatusCode <= 399)
-        {
-            return ActivityStatusCode.Unset;
-        }
-
-        if (httpStatusCode == 404)
-        {
-            return ActivityStatusCode.Unset;
-        }
-
-        return ActivityStatusCode.Error;
-    }
-
-    /// <summary>
     /// Helper method that populates Activity Status from http status code according
-    /// to https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/http.md#status.
+    /// to https://github.com/open-telemetry/semantic-conventions/blob/v1.24.0/docs/http/http-spans.md#status.
     /// </summary>
     /// <param name="kind">The span kind.</param>
     /// <param name="httpStatusCode">Http status code.</param>